[EGD-7610] Unnecessary services and apps removed
Removed lwIP, FOTA and clock application.
