+++ CMakeLists.txt
@@ -18,7 +18,6 @@ include(Colours)
 include(CCacheConfig)
 include(ProjectConfig)
 include(ModuleConfig)
-include(module-lwip/lwip-includes.cmake)
 include(SerialPort)
 include(CopyGdbInit)
 include(Utils)
@@ -131,7 +130,6 @@ add_subdirectory(module-audio)
 add_subdirectory(module-services)
 add_subdirectory(module-apps)
 add_subdirectory(module-bluetooth)
-add_subdirectory(module-lwip)
 add_subdirectory(image)
 add_subdirectory(host-tools)
