--- module-lwip/lib/dhcp-server/README
@@ -1,2 +1,0 @@
-DHCP Server for lwIP from stm32 ethernet over usb (rndis + lwip) project
-https://github.com/fetisov/lrndis
