+++ module-services/service-bluetooth/service-bluetooth/BluetoothMessage.hpp
@@ -35,7 +35,6 @@ class BluetoothMessage : public sys::DataMessage
         Scan,
         StopScan,
         getDevicesAvailable,
-        PAN,
         Visible,
         Play,
         SwitchProfile,
