+++ products/PurePhone/CMakeLists.txt
@@ -69,10 +69,8 @@ target_link_libraries(PurePhone
         service-bluetooth
         service-desktop
         service-fileindexer
-        service-lwip
         service-time
         sys
-        ${LWIP_LIBRARIES}
         "$<$<STREQUAL:${PROJECT_TARGET},TARGET_Linux>:iosyscalls>"
         "$<$<STREQUAL:${PROJECT_TARGET},TARGET_RT1051>:CrashCatcher::CrashCatcher>"
     )
