+++ products/PurePhone/PurePhoneMain.cpp
@@ -43,7 +43,6 @@
 #include <service-audio/ServiceAudio.hpp>
 #include <service-bluetooth/ServiceBluetooth.hpp>
 #include <service-db/agents/quotes/QuotesAgent.hpp>
-#include <service-lwip/ServiceLwIP.hpp>
 #include <service-time/ServiceTime.hpp>
 #include <Service/ServiceCreator.hpp>
 #include <service-gui/ServiceGUI.hpp>
@@ -53,7 +52,6 @@
 #include <service-desktop/ServiceDesktop.hpp>
 
 #if ENABLE_GSM == 1
-#include <service-fota/ServiceFota.hpp>
 #include <service-cellular/ServiceCellular.hpp>
 #include <service-antenna/ServiceAntenna.hpp>
 #endif
@@ -103,11 +101,9 @@ int main()
 #else
     systemServices.emplace_back(sys::CreatorFor<ServiceAntenna>());
     systemServices.emplace_back(sys::CreatorFor<ServiceCellular>());
-    systemServices.emplace_back(sys::CreatorFor<FotaService::Service>());
 #endif
     systemServices.emplace_back(sys::CreatorFor<ServiceAudio>());
     systemServices.emplace_back(sys::CreatorFor<ServiceBluetooth>());
-    systemServices.emplace_back(sys::CreatorFor<ServiceLwIP>());
     systemServices.emplace_back(sys::CreatorFor<ServiceDesktop>());
     systemServices.emplace_back(sys::CreatorFor<stm::ServiceTime>());
     systemServices.emplace_back(sys::CreatorFor<service::eink::ServiceEink>());
